Migrate TestGetFirefoxButton.js
